# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pixel WEB SERVER",
    "version": "1.0.0",
    "description": "Документація до беку"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "https",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
...
    "/users/profile": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "User"
        ],
        "summary": "Профіль користувача",
        "description": "Цей запит дозволяє отримати інформацію про профіль користувача, який автентифікований за допомогою JWT токена. Необхідно надати заголовок `Authorization` з дійсним JWT токеном.",
        "responses": {
          "200": {
            "description": "Успішна відповідь з інформацією про користувача",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "_id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "fullname": {
                          "type": "string"
                        },
                        "phone": {
                          "type": "string"
                        },
                        "history": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "favorite": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "deleted": {
                          "type": "boolean"
                        }
                      }
                    },
                    "accessToken": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "user": {
                    "_id": "66e0754c092b2c7a8eea094d",
                    "email": "ivan@gmaill.com",
                    "fullname": "Мийхай Іван Миколайович",
                    "phone": "0937226252",
                    "history": [],
                    "favorite": [],
                    "createdAt": "2024-09-10T16:35:19.023Z"
                  },
                  "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
                }
              }
            }
          },
          "401": {
            "description": "Неправильний або відсутній токен авторизації",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Unauthorized"
                }
              }
            }
          },
          "500": {
            "description": "Помилка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
    "/users/history": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "User"
        ],
        "summary": "Отримати історію замовлень користувача",
        "description": "Цей запит дозволяє отримати список всіх замовлень користувача, які не були видалені. Необхідно надати заголовок `Authorization` з дійсним JWT токеном.",
        "responses": {
          "200": {
            "description": "Успішна відповідь зі списком замовлень користувача",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orders": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string"
                          },
                          "orderNumber": {
                            "type": "string"
                          },
                          "localStorage": {
                            "type": "array",
                            "items": {
                              "type": "object"
                            }
                          },
                          "cost": {
                            "type": "number"
                          },
                          "fullname": {
                            "type": "string"
                          },
                          "phone": {
                            "type": "string"
                          },
                          "deliveryType": {
                            "type": "string"
                          },
                          "city": {
                            "type": "string"
                          },
                          "address": {
                            "type": "string"
                          },
                          "customerComment": {
                            "type": "string"
                          },
                          "managerComment": {
                            "type": "string"
                          },
                          "invoiceId": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "deleted": {
                            "type": "boolean"
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "orders": [
                    {
                      "_id": "64e0754c092b2c7a8eea094d",
                      "orderNumber": "ORD123456",
                      "localStorage": [
                        {
                          "goodsId": "prod1",
                          "quantity": 2
                        }
                      ],
                      "cost": 400.00,
                      "fullname": "Іван Миколайович",
                      "phone": "0937226252",
                      "deliveryType": "Courier",
                      "city": "Kyiv",
                      "address": "123 Main St",
                      "customerComment": "Please deliver after 6 PM",
                      "managerComment": "Urgent delivery",
                      "invoiceId": "inv123456",
                      "createdAt": "2024-09-10T16:35:19.023Z",
                      "deleted": false
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Замовлення не знайдено для цього користувача",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "No orders found for this user"
                }
              }
            }
          },
          "500": {
            "description": "Помилка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
    "/users/favorite": {
      "get": {
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "tags": [
          "User"
        ],
        "summary": "Отримати список улюблених",
        "description": "Цей запит дозволяє отримати список улюблених продуктів користувача. Необхідно надати заголовок `Authorization` з дійсним JWT токеном.",
        "responses": {
          "200": {
            "description": "Успішна відповідь зі списком улюблених продуктів користувача",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "favorite": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "_id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "price": {
                            "type": "number"
                          },
                          "description": {
                            "type": "string"
                          },
                          "photo": {
                            "type": "string"
                          },
                          "article": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                },
                "example": {
                  "favorite": [
                    {
                      "_id": "64e0754c092b2c7a8eea094d",
                      "name": "Продукт 1",
                      "price": 100.00,
                      "description": "Опис продукту 1",
                      "photo": "http://example.com/photo1.jpg",
                      "article": "ART123",
                      "createdAt": "2024-09-10T16:35:19.023Z"
                    },
                    {
                      "_id": "64e0754c092b2c7a8eea094e",
                      "name": "Продукт 2",
                      "price": 150.00,
                      "description": "Опис продукту 2",
                      "photo": "http://example.com/photo2.jpg",
                      "article": "ART124",
                      "createdAt": "2024-09-10T16:35:19.023Z"
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Улюблені продукти не знайдені",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "message": "No favorite found"
                }
              }
            }
          },
          "500": {
            "description": "Помилка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                },
                "example": {
                  "error": "Internal server error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Додати товари до улюблених користувача",
        "description": "Цей ендпоінт дозволяє користувачам додавати товари до списку улюблених.",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "goods": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "objectId"
                    }
                  }
                },
                "example": {
                  "goods": [
                    "60d21b4667d0d8992e610c85",
                    "60d21b4667d0d8992e610c86"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Товари успішно додані до улюблених",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Успішно додано"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Товари не знайдено",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Товари не знайдено"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Внутрішня помилка сервера"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Оновлення профілю користувача",
        "description": "Цей ендпоінт дозволяє користувачам оновлювати свій профіль, включаючи номер телефону, місто та адресу.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string",
                    "description": "Новий номер телефону"
                  },
                  "city": {
                    "type": "string",
                    "description": "Нове місто"
                  },
                  "address": {
                    "type": "string",
                    "description": "Нова адреса"
                  }
                },
                "example": {
                  "phone": "0951234567",
                  "city": "Київ",
                  "address": "Вулиця Шевченка, 1"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Профіль успішно оновлено",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Profile updated successfully"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "email": {
                          "type": "string"
                        },
                        "fullname": {
                          "type": "string"
                        },
                        "phone": {
                          "type": "string"
                        },
                        "city": {
                          "type": "string"
                        },
                        "address": {
                          "type": "string"
                        },
                        "history": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "format": "objectId"
                          }
                        },
                        "favorite": {
                          "type": "array",
                          "items": {
                            "type": "string",
                            "format": "objectId"
                          }
                        }
                      }
                    },
                    "accessToken": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Видалення товарів з обраного списку",
        "description": "Цей ендпоінт дозволяє користувачам видаляти товари зі свого обраного списку.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Список ідентифікаторів товарів, які потрібно видалити, розділених комами.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "60d21b4667d0d8992e610c85,60d21b4667d0d8992e610c86"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Товари успішно видалено з обраного списку",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Deleted from favorite"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/orders": {
      "delete": {
        "tags": [
          "User"
        ],
        "summary": "Видалення замовлень користувача",
        "description": "Цей ендпоінт дозволяє користувачам видаляти свої замовлення. Можна видалити всі замовлення або вибрані за ідентифікаторами.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Список ідентифікаторів замовлень, які потрібно видалити, розділених комами. Якщо не вказано, будуть видалені всі замовлення користувача.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "60d21b4667d0d8992e610c85,60d21b4667d0d8992e610c86"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Замовлення успішно видалено",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "All Orders deleted"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/review": {
      "post": {
        "summary": "Додати відгук до товару",
        "description": "Цей ендпоінт дозволяє користувачам додавати відгуки до товарів.",
        "tags": [
          "User"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "ID товару, до якого додається відгук",
            "required": true,
            "schema": {
              "type": "string",
              "format": "objectId"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {
                    "type": "string",
                    "description": "Текст відгуку"
                  }
                },
                "example": {
                  "text": "Чудовий товар, рекомендую!"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Відгук успішно створено",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Successfully created"
                    },
                    "review": {
                      "type": "object",
                      "properties": {
                        "reviewSenderId": {
                          "type": "string",
                          "format": "objectId"
                        },
                        "product": {
                          "type": "string",
                          "format": "objectId"
                        },
                        "content": {
                          "type": "object",
                          "properties": {
                            "text": {
                              "type": "string"
                            }
                          }
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Не вдалося створити відгук",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Review not created"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Внутрішня помилка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}